benchmarker/requirements.txt
pymilvus==2.5.6
pandas==2.2.3
pyarrow==18.1.0
psutil==6.1.1
scikit-learn==1.6.0
weaviate-client==4.10.2
SQLAlchemy==2.0.37
psycopg2-binary==2.9.10
pgvector==0.3.6
qdrant-client==1.13.2
elasticsearch==8.17.1
chromadb==0.6.3